feat(FE: Reducers): Combine Metrics reducers and refactor Metrics actions
